style: 팝업박스 리모델링
